functions/requirements.txt
firebase-admin
flask
requests
beautifulsoup4
sendgrid
google-cloud-pubsub
google-cloud-aiplatform
lxml
